requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter ecosystem
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# LangChain and LLM libraries
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5

# Data validation and parsing
pydantic>=2.0.0

# Token counting
tiktoken>=0.5.0

# Environment management
python-dotenv>=1.0.0

# MongoDB database
pymongo>=4.6.0
